requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.0.1
PyPDF2==3.0.1
pydantic==2.5.1
python-jose==3.3.0
tabula-py==2.9.0
aiofiles==23.2.1
jinja2==3.1.2 